refactor: estrarre il motore pipeline da usePipeline.ts
usePipeline.ts era un motore monolitico (stage execution, retry, judge,
stato run) accoppiato a React solo tramite useCallback e useTranslation,
ma già scritto in stile imperativo (chunksStore/pipelineStore via
getState() quasi ovunque). Il bridge trascrizione→traduzione (#224)
dovrà riusare la stessa orchestrazione da un entry point che non parte
da un componente React, quindi va separata prima che #224 la riusi.

Sposta stage execution/retry/judge/blob/persistenza stato in
hooks/pipeline/engine.ts: funzioni pure, nessun import React, la
traduzione (t) è iniettata come parametro anziché letta da
useTranslation(). usePipeline.ts diventa un wiring sottile che inietta
t e riespone la stessa API pubblica (runPipeline, runSingleChunk,
rerunChunkWithMemory, cancelPipeline, ecc.) — nessun cambio di
comportamento, verificato dalla suite di test esistente (676 test,
invariata) e da typecheck/lint puliti.

Come effetto collaterale risolve la nota lasciata nel codice originale
su executePipelineForChunk ridefinita a ogni render (non più un
problema: ora è una funzione a livello di modulo, non ricreata).
